Rewrite HTTP/2 responses as HTTP1.1
Some platforms don't route HTTP/2, but some providers have trouble with
HTTP/1.1 requests. This changes the proxied request to use HTTP/whatever and
rewrites the response to always be HTTP/1.1.

It also drops some unnecessary echo calls, introduces a logging
function, a HTTP status checker and logic to return a 502 if it
encounters anything but OK.